+++ doc/CHANGES.rst
@@ -23,6 +23,7 @@ CHANGELOG
 * [FIX] Field.DateTime documentation, add ``is auto updated``
 * [REF] add entry point ``anyblok.session.event`` and additional_setting 
   ``anyblok.session.event`` to add some events on the session
+* [FIX] clean foreign_key in some column type, now the foreign_key is made by Column class
 
 0.11.1 (2017-11-28)
 -------------------